feat(tools): add execution parameter to create/update deferred prompt tools
Extract handler functions to tool-handlers.ts to stay within max-lines limit.
Move shared schemas and parseExecutionMetadata to types.ts.
